Loose the matrix size from 4*4 to M*N
Modify the functions of naive version to handle the situations that
the matrices's size differ from 4*4. Also, I use random function to
form the two source matrices, so that it would be more realistic.

**WARNING**
I still can't find some effective method to prove that the product
of the matrix mutiplication is correct, but I have manually checked
some cases and the answer is correct.